apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
